chore: Update tapir to latest version
Remove index page and duplicate /health endpoint
