+++ webapi/src/main/scala/org/knora/webapi/slice/infrastructure/api/PrometheusApp.scala
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.knora.webapi.instrumentation.prometheus
+package org.knora.webapi.slice.infrastructure.api
 
 import zio._
 import zio.http._
@@ -12,14 +12,9 @@ import zio.metrics.connectors.prometheus.PrometheusPublisher
 /**
  * Provides the '/metrics' endpoint serving the metrics in prometheus format.
  */
-final case class PrometheusApp() {
-
-  val route: HttpApp[PrometheusPublisher, Nothing] =
-    Http
-      .collectZIO[Request] { case Method.GET -> Root / "metrics" =>
-        ZIO.serviceWithZIO[PrometheusPublisher](_.get.map(Response.text))
-      }
+final case class PrometheusApp(prometheus: PrometheusPublisher) {
+  val route: HttpApp[Any] = Routes(Method.GET / "metrics" -> handler(prometheus.get.map(Response.text(_)))).toHttpApp
 }
 object PrometheusApp {
-  val layer = ZLayer.succeed(PrometheusApp())
+  val layer = ZLayer.derive[PrometheusApp]
 }
